StorageConverterStorageValueToStorageBinarySharedXXX was StorageConverterToStorageBinarySharedXXX
